+++ src/modules/storefront/interfaces/index.ts
@@ -1,14 +1,27 @@
 export interface TemplateData {
   title: string;
   items: {
-    type: ComponentType;
-    props: HeaderData | BannerData;
+    type: keyof Template;
+    props: HeaderData | BannerData | HeroData | MenuData;
   }[];
 }
+
+export type DefaultDataProps =
+  | HeaderProps
+  | BannerProps
+  | HeroProps
+  | MenuProps;
+
+type HeaderProps = { data: HeaderData; defaultData: HeaderDefault };
+type BannerProps = { data: BannerData; defaultData: BannerDefault };
+type MenuProps = { data: MenuData; defaultData: MenuDefault };
+type HeroProps = { data: HeroData; defaultData: HeroDefault };
 
 export interface Template {
   header: HeaderDefault;
   banner: BannerDefault;
+  menu: MenuDefault;
+  hero: HeroDefault;
 }
 
 export type HeaderData = {
@@ -28,10 +41,29 @@ export type BannerData = {
   textColor?: string;
   categories?: CategoryItem[];
 };
+export type MenuData = {
+  bgColor?: string;
+  textColor?: string;
+  categories?: CategoryItem[];
+};
+
+export type MenuDefault = {
+  bgColor: string;
+  textColor: string;
+  categories?: CategoryItem[];
+};
 
 export type BannerDefault = {
   bgColor: string;
   textColor: string;
+};
+
+export type HeroData = {
+  data?: string;
+};
+
+export type HeroDefault = {
+  data: string;
 };
 
 type CategoryItem = { label: string; slug: string };
@@ -41,5 +73,3 @@ type HeaderLink = {
   value: string;
   newWindow: boolean;
 };
-
-type ComponentType = 'header' | 'banner';